Added ImGUI support
My plan was to use it as main gui but after working with it a little bit this will be debug only gui solution. Wraper will be extended since now its preety basic, and i thing it will be only included in debug builds but it will be implemented in the future.

TLDR: it dirty, simple, basic, unfinished implementation of imgui