Engine: Added proper CRT effect
It might be too much for some games, but as always its stylistic choice. I would highly recomend toggle for this effect in your game since it looks bes on high res displays.

Yes im bored and making shaders instead of fixing countless thing broken or done wrong in the engine. It is what it is.